+++ src/AssemblyInfo.cs
@@ -1,0 +1,18 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+using IonicPack;
+
+[assembly: AssemblyTitle(VSPackage.Name)]
+[assembly: AssemblyDescription("An extension to make you more productive working with the Ionic Framework")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct(VSPackage.Name)]
+[assembly: AssemblyCopyright("Copyright ©  2016")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+[assembly: ComVisible(false)]
+
+[assembly: Guid(VSPackage.Guid)]
+
+[assembly: AssemblyVersion(VSPackage.Version)]
+[assembly: AssemblyFileVersion(VSPackage.Version)]