+++ src/HTML/Schema/HtmlCache.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using Newtonsoft.Json;
@@ -11,12 +12,20 @@ namespace IonicPack.HTML
 
         static List<HtmlElement> ReadConfig()
         {
-            string assembly = Assembly.GetExecutingAssembly().Location;
-            string folder = Path.GetDirectoryName(assembly);
-            string source = Path.Combine(folder, "HTML\\Schema\\schema.json");
+            try
+            {
+                string assembly = Assembly.GetExecutingAssembly().Location;
+                string folder = Path.GetDirectoryName(assembly);
+                string source = Path.Combine(folder, "HTML\\Schema\\schema.json");
 
-            string content = File.ReadAllText(source);
-            return JsonConvert.DeserializeObject<List<HtmlElement>>(content);
+                string content = File.ReadAllText(source);
+                return JsonConvert.DeserializeObject<List<HtmlElement>>(content);
+            }
+            catch (Exception ex)
+            {
+                Logger.Log(ex);
+                return new List<HtmlElement>();
+            }
         }
     }
 }